requirements.txt
tinygrad>=0.9.0
tokenizers>=0.15.0
PyYAML>=6.0.1
numpy>=1.24.0
